+++ src/Context/MapperOptions.php
@@ -19,12 +19,14 @@ namespace Rekalogika\Mapper\Context;
 final readonly class MapperOptions
 {
     /**
-     * @param boolean $lazyLoading Enable or disable lazy loading.
+     * @param boolean $lazyLoading Enable lazy loading.
      * @param boolean $readTargetValue If disabled, values on the target side will not be read, and assumed to be null.
+     * @param boolean $objectToObjectScalarShortCircuit Performance optimization by doing scalar to scalar transformation within `ObjectToObjectTransformer` instead of delegating to the `MainTransformer`
      */
     public function __construct(
         public bool $lazyLoading = true,
         public bool $readTargetValue = true,
+        public bool $objectToObjectScalarShortCircuit = true,
     ) {
     }
 }